+++ resources/views/components/layouts/app.blade.php
@@ -5,6 +5,13 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>{{ $title ?? env('APP_NAME') }}</title>
+    
+    <link
+        rel="alternate"
+        type="application/rss+xml"
+        title="Artemis Journal RSS"
+        href="{{ route('feeds.main') }}" />
+
 
     {!! $meta ?? '<!-- page default meta -->' !!}